플레이어 경험치 테이블 파싱 구현, 파티클 오류 수정 중
- json을 활용하여 플레이어 경험치 테이블 파싱 구현
- 파티클 시스템 오류 수정 중
  : 몬스터가 살아있는데도 파티클 시스템이 제거되는 현상

+++ Spirit Strike/Assets/Scripts/PlayerData.cs
@@ -11,5 +11,5 @@ public class PlayerData
     public int _defence;
     public int _dodge;
     public int _critical;
-    public float _atkSpeed;
+    public int _atkSpeed;
 }
